rename TrimSnapshot
