+++ src/server/graphql/models/Organization/addBilling/__tests__/addBilling.test.js
@@ -4,8 +4,8 @@ import mockAuthToken from 'server/__tests__/setup/mockAuthToken';
 import stripe from 'server/billing/stripe';
 import MockDate from 'mockdate';
 import {__now} from 'server/__tests__/setup/mockTimes';
-import fetchAndTrim from 'server/__tests__/utils/fetchAndTrim';
-import TrimSnapshot from 'server/__tests__/utils/TrimSnapshot';
+import fetchAndSerialize from 'server/__tests__/utils/fetchAndSerialize';
+import DynamicSerializer from 'server/__tests__/utils/DynamicSerializer';
 import MockDB from 'server/__tests__/setup/MockDB';
 import {PAYMENT_REJECTED, TRIAL_EXPIRES_SOON, TRIAL_EXPIRED} from 'universal/utils/constants';
 import creditCardByToken from 'server/__tests__/utils/creditCardByToken';
@@ -20,12 +20,12 @@ describe('addBilling', () => {
   test('extends trial for those still in trial period', async() => {
     // SETUP
     const r = getRethink();
-    const trimSnapshot = new TrimSnapshot();
+    const dynamicSerializer = new DynamicSerializer();
     const mockDB = new MockDB();
     const {user, organization} = await mockDB.init()
       .newNotification(undefined, {type: TRIAL_EXPIRES_SOON});
     const org = organization[0];
-    stripe.__setMockData(org, trimSnapshot);
+    stripe.__setMockData(org, dynamicSerializer);
     const stripeToken = 'tok_4242424242424242';
     const orgId = org.id;
     const authToken = mockAuthToken(user[0]);
@@ -34,10 +34,10 @@ describe('addBilling', () => {
     await addBilling.resolve(undefined, {orgId, stripeToken}, {authToken, socket});
 
     // VERIFY
-    const db = await fetchAndTrim({
+    const db = await fetchAndSerialize({
       organization: r.table('Organization').get(orgId),
       notification: r.table('Notification').getAll(orgId, {index: 'orgId'})
-    }, trimSnapshot);
+    }, dynamicSerializer);
     expect(db).toMatchSnapshot();
     expect(stripe.__snapshot()).toMatchSnapshot();
   });
@@ -45,7 +45,7 @@ describe('addBilling', () => {
   test('starts a new subscription for those who let the trial expire', async() => {
     // SETUP
     const r = getRethink();
-    const trimSnapshot = new TrimSnapshot();
+    const dynamicSerializer = new DynamicSerializer();
     const mockDB = new MockDB();
     const stripeToken = 'tok_4242424242424242';
     const {user, team, organization} = await mockDB.init()
@@ -53,7 +53,7 @@ describe('addBilling', () => {
       .org(0, {periodEnd: new Date(now.getTime() - 1)})
       .team(0, {isPaid: false});
     const org = organization[0];
-    stripe.__setMockData(org, trimSnapshot);
+    stripe.__setMockData(org, dynamicSerializer);
     const orgId = org.id;
     const authToken = mockAuthToken(user[0]);
 
@@ -61,11 +61,11 @@ describe('addBilling', () => {
     await addBilling.resolve(undefined, {orgId, stripeToken}, {authToken, socket});
 
     // VERIFY
-    const db = await fetchAndTrim({
+    const db = await fetchAndSerialize({
       organization: r.table('Organization').get(orgId),
       notification: r.table('Notification').getAll(orgId, {index: 'orgId'}),
       team: r.table('Team').get(team[0].id)
-    }, trimSnapshot);
+    }, dynamicSerializer);
     expect(db).toMatchSnapshot();
     expect(stripe.__snapshot()).toMatchSnapshot();
   });
@@ -73,13 +73,13 @@ describe('addBilling', () => {
   test('changes cards for customers in good standing', async() => {
     // SETUP
     const r = getRethink();
-    const trimSnapshot = new TrimSnapshot();
+    const dynamicSerializer = new DynamicSerializer();
     const mockDB = new MockDB();
     const oldToken = 'tok_4012888888881881';
     const {user, organization} = await mockDB.init()
       .org(0, {creditCard: creditCardByToken[oldToken]});
     const org = organization[0];
-    stripe.__setMockData(org, trimSnapshot);
+    stripe.__setMockData(org, dynamicSerializer);
     const orgId = org.id;
     const authToken = mockAuthToken(user[0]);
 
@@ -88,25 +88,23 @@ describe('addBilling', () => {
     await addBilling.resolve(undefined, {orgId, stripeToken}, {authToken, socket});
 
     // VERIFY
-    const db = await fetchAndTrim({
+    const db = await fetchAndSerialize({
       organization: r.table('Organization').get(orgId),
-      // notification: r.table('Notification').getAll(orgId, {index: 'orgId'}),
-      // team: r.table('Team').get(team[0].id)
-    }, trimSnapshot);
+    }, dynamicSerializer);
     expect(db).toMatchSnapshot();
-    expect(stripe.__snapshot()).toMatchSnapshot();
+    expect(stripe.__snapshot(dynamicSerializer)).toMatchSnapshot();
   });
 
   test('changes cards for customers in good standing', async() => {
     // SETUP
     const r = getRethink();
-    const trimSnapshot = new TrimSnapshot();
+    const dynamicSerializer = new DynamicSerializer();
     const mockDB = new MockDB();
     const oldToken = 'tok_4012888888881881';
     const {user, organization} = await mockDB.init()
       .org(0, {creditCard: creditCardByToken[oldToken]});
     const org = organization[0];
-    stripe.__setMockData(org, trimSnapshot);
+    stripe.__setMockData(org, dynamicSerializer);
     const orgId = org.id;
     const authToken = mockAuthToken(user[0]);
 
@@ -115,11 +113,11 @@ describe('addBilling', () => {
     await addBilling.resolve(undefined, {orgId, stripeToken}, {authToken, socket});
 
     // VERIFY
-    const db = await fetchAndTrim({
+    const db = await fetchAndSerialize({
       organization: r.table('Organization').get(orgId),
       // notification: r.table('Notification').getAll(orgId, {index: 'orgId'}),
       // team: r.table('Team').get(team[0].id)
-    }, trimSnapshot);
+    }, dynamicSerializer);
     expect(db).toMatchSnapshot();
     expect(stripe.__snapshot()).toMatchSnapshot();
   });
@@ -127,7 +125,7 @@ describe('addBilling', () => {
   test('changes cards for customer who had a failed charge', async() => {
     // SETUP
     const r = getRethink();
-    const trimSnapshot = new TrimSnapshot();
+    const dynamicSerializer = new DynamicSerializer();
     const mockDB = new MockDB();
     const oldToken = 'tok_4000000000000341';
     const {user, team, organization} = await mockDB.init()
@@ -137,7 +135,7 @@ describe('addBilling', () => {
       })
       .newNotification(undefined, {type: PAYMENT_REJECTED});
     const org = organization[0];
-    stripe.__setMockData(org, trimSnapshot);
+    stripe.__setMockData(org, dynamicSerializer);
     const orgId = org.id;
     const authToken = mockAuthToken(user[0]);
 
@@ -146,11 +144,11 @@ describe('addBilling', () => {
     await addBilling.resolve(undefined, {orgId, stripeToken}, {authToken, socket});
 
     // VERIFY
-    const db = await fetchAndTrim({
+    const db = await fetchAndSerialize({
       organization: r.table('Organization').get(orgId),
       notification: r.table('Notification').getAll(orgId, {index: 'orgId'}),
       team: r.table('Team').get(team[0].id)
-    }, trimSnapshot);
+    }, dynamicSerializer);
     expect(db).toMatchSnapshot();
     expect(stripe.__snapshot()).toMatchSnapshot();
   });
