+++ src/server/graphql/models/Organization/addOrg/__tests__/addOrg.test.js
@@ -4,8 +4,8 @@ import mockAuthToken from 'server/__tests__/setup/mockAuthToken';
 import stripe from 'server/billing/stripe';
 import MockDate from 'mockdate';
 import {__now} from 'server/__tests__/setup/mockTimes';
-import fetchAndTrim from 'server/__tests__/utils/fetchAndTrim';
-import TrimSnapshot from 'server/__tests__/utils/TrimSnapshot';
+import fetchAndSerialize from 'server/__tests__/utils/fetchAndSerialize';
+import DynamicSerializer from 'server/__tests__/utils/DynamicSerializer';
 import MockDB from 'server/__tests__/setup/MockDB';
 import creditCardByToken from 'server/__tests__/utils/creditCardByToken';
 import socket from 'server/__mocks__/socket';
@@ -19,13 +19,13 @@ describe('addOrg', () => {
   test('adds a new org with no invitees', async() => {
     // SETUP
     const r = getRethink();
-    const trimSnapshot = new TrimSnapshot();
+    const dynamicSerializer = new DynamicSerializer();
     const mockDB = new MockDB();
     const stripeToken = 'tok_4242424242424242';
     const {organization, user} = await mockDB.init()
       .org(0, {creditCard: creditCardByToken[stripeToken]});
     const org = organization[0];
-    stripe.__setMockData(org, trimSnapshot);
+    stripe.__setMockData(org, dynamicSerializer);
     auth0ManagementClient.__initMock(mockDB.db);
     const authToken = mockAuthToken(user[0]);
 
@@ -39,12 +39,12 @@ describe('addOrg', () => {
     await addOrg.resolve(undefined, {newTeam, orgName, stripeToken}, {authToken, socket});
 
     // VERIFY
-    const db = await fetchAndTrim({
+    const db = await fetchAndSerialize({
       organization: r.table('Organization').getAll(newTeam.orgId, {index: 'id'}).orderBy('name'),
       team: r.table('Team').getAll(newTeam.orgId, {index: 'orgId'}).orderBy('name'),
       teamMember: r.table('TeamMember').getAll(newTeam.id, {index: 'teamId'}).orderBy('preferredName'),
       user: r.table('User').getAll(org.id, newTeam.orgId, {index: 'userOrgs'}).orderBy('preferredName')
-    }, trimSnapshot);
+    }, dynamicSerializer);
     expect(db).toMatchSnapshot();
     expect(stripe.__snapshot()).toMatchSnapshot();
   });
