+++ src/server/graphql/models/Team/endMeeting/__tests__/endMeeting.test.js
@@ -2,8 +2,8 @@ import getRethink from 'server/database/rethinkDriver';
 import mockAuthToken from 'server/__tests__/setup/mockAuthToken';
 import MockDate from 'mockdate';
 import {__now} from 'server/__tests__/setup/mockTimes';
-import fetchAndTrim from 'server/__tests__/utils/fetchAndTrim';
-import TrimSnapshot from 'server/__tests__/utils/TrimSnapshot';
+import fetchAndSerialize from 'server/__tests__/utils/fetchAndSerialize';
+import DynamicSerializer from 'server/__tests__/utils/DynamicSerializer';
 import MockDB from 'server/__tests__/setup/MockDB';
 import expectAsyncToThrow from 'server/__tests__/utils/expectAsyncToThrow';
 import socket from 'server/__mocks__/socket';
@@ -17,7 +17,7 @@ describe('endMeeting', () => {
   test('generates a meeting summary and sets sort order', async() => {
     // SETUP
     const r = getRethink();
-    const trimSnapshot = new TrimSnapshot();
+    const dynamicSerializer = new DynamicSerializer();
     const mockDB = new MockDB();
     const {teamMember, user} = await mockDB.init()
       .newMeeting(undefined, {inProgress: true});
@@ -30,14 +30,14 @@ describe('endMeeting', () => {
     // TEST
     await endMeeting.resolve(undefined, {teamId}, {authToken, socket});
     // VERIFY
-    const db = await fetchAndTrim({
+    const db = await fetchAndSerialize({
       agendaItem: r.table('AgendaItem').getAll(teamId, {index: 'teamId'}).orderBy('teamMemberId'),
       action: r.table('Action').getAll(r.args(teamMemberIds), {index: 'teamMemberId'}).orderBy('teamMemberId'),
       project: r.table('Project').getAll(r.args(teamMemberIds), {index: 'teamMemberId'}).orderBy('teamMemberId'),
       meeting: r.table('Meeting').get(meetingId),
       team: r.table('Team').get(teamId),
       teamMember: r.table('TeamMember').getAll(teamId, {index: 'teamId'}).orderBy('preferredName'),
-    }, trimSnapshot);
+    }, dynamicSerializer);
     expect(db).toMatchSnapshot();
     expect(mockFn).toBeCalledWith(teamId);
   });
@@ -45,7 +45,7 @@ describe('endMeeting', () => {
   test('generates a meeting summary and sets sort order with pre-existing actions and projects', async() => {
     // SETUP
     const r = getRethink();
-    const trimSnapshot = new TrimSnapshot();
+    const dynamicSerializer = new DynamicSerializer();
     const mockDB = new MockDB();
     const {teamMember, user} = await mockDB.init()
       .newAction()
@@ -60,14 +60,14 @@ describe('endMeeting', () => {
     // TEST
     await endMeeting.resolve(undefined, {teamId}, {authToken, socket});
     // VERIFY
-    const db = await fetchAndTrim({
+    const db = await fetchAndSerialize({
       agendaItem: r.table('AgendaItem').getAll(teamId, {index: 'teamId'}).orderBy('content'),
       action: r.table('Action').getAll(r.args(teamMemberIds), {index: 'teamMemberId'}).orderBy('content'),
       project: r.table('Project').getAll(r.args(teamMemberIds), {index: 'teamMemberId'}).orderBy('content'),
       meeting: r.table('Meeting').get(meetingId),
       team: r.table('Team').get(teamId),
       teamMember: r.table('TeamMember').getAll(teamId, {index: 'teamId'}).orderBy('preferredName'),
-    }, trimSnapshot);
+    }, dynamicSerializer);
     expect(db).toMatchSnapshot();
     expect(mockFn).toBeCalledWith(teamId);
   });
